Separate psychometric export tests
