+++ apps/backend/src/app/database/services/coding/coding-psychometric-export.service.spec.ts
@@ -1,1079 +1,166 @@
-import * as fs from 'fs';
 import { PassThrough } from 'stream';
 import { CodingPsychometricExportService } from './coding-psychometric-export.service';
-
-interface AnalysisResponseRow {
-  responseId: number;
-  personId: number;
-  unitName: string;
-  variableId: string;
-  value: string;
-  statusV1: number;
-  codeV1: number;
-  scoreV1: number;
-  statusV2: number;
-  codeV2: number;
-  scoreV2: number;
-  statusV3: number;
-  codeV3: number;
-  scoreV3: number;
-}
+import { PsychometricAnalysis } from './psychometric-export.types';
 
 describe('CodingPsychometricExportService', () => {
-  const createService = (
-    vomd: Record<string, unknown>,
-    variables: Array<Record<string, unknown>>
-  ) => new CodingPsychometricExportService(
-    {
-      find: jest.fn().mockResolvedValue([
-        {
-          id: 1,
-          file_id: 'UNIT_A.VOMD',
-          filename: 'UNIT_A.vomd',
-          data: JSON.stringify(vomd)
-        }
-      ])
-    } as never,
-    {
-      getUnitVariableDetails: jest.fn().mockResolvedValue([
-        {
-          unitName: 'UNIT_A',
-          unitId: 'UNIT_A',
-          variables
-        }
-      ])
-    } as never,
-    {} as never,
-    {} as never,
-    {} as never
-  );
+  const createService = () => {
+    const item = {
+      key: 'UNIT_A\u001FV1',
+      unitName: 'UNIT_A',
+      variableId: 'V1'
+    };
+    const mapping = {
+      items: [item],
+      byLogicalKey: new Map([[item.key, item]]),
+      issues: []
+    };
+    const snapshot = {
+      duplicatePersonIds: new Set<number>(),
+      totalRows: 0,
+      forEachBatch: jest.fn()
+    };
+    const analysis: PsychometricAnalysis = {
+      rows: [],
+      summary: []
+    };
+    const metadataResolver = {
+      getDomainCandidates: jest.fn().mockResolvedValue({
+        candidates: [],
+        mappingIssueCount: 0
+      }),
+      buildItemMapping: jest.fn().mockResolvedValue(mapping),
+      assignDomains: jest.fn(),
+      loadMissingDefinitions: jest.fn().mockResolvedValue([])
+    };
+    const responseReader = {
+      withSnapshot: jest.fn(
+        async (
+          _input: unknown,
+          callback: (value: typeof snapshot) => Promise<unknown>
+        ) => callback(snapshot)
+      )
+    };
+    const analysisEngine = {
+      analyze: jest.fn().mockResolvedValue(analysis)
+    };
+    const exportWriter = {
+      createCsvStream: jest.fn().mockReturnValue(new PassThrough()),
+      writeExcelToFile: jest.fn().mockResolvedValue(undefined)
+    };
+    const service = new CodingPsychometricExportService(
+      metadataResolver as never,
+      responseReader as never,
+      analysisEngine as never,
+      exportWriter as never
+    );
 
-  const collectStream = async (
-    stream: NodeJS.ReadableStream
-  ): Promise<string> => {
-    const chunks: Buffer[] = [];
-    for await (const chunk of stream as AsyncIterable<Buffer | string>) {
-      chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-    }
-    return Buffer.concat(chunks).toString('utf8');
+    return {
+      service,
+      mapping,
+      snapshot,
+      analysis,
+      metadataResolver,
+      responseReader,
+      analysisEngine,
+      exportWriter
+    };
   };
 
-  const createAnalysisService = (
-    responseOverrides: Record<number, Partial<AnalysisResponseRow>> = {}
-  ): CodingPsychometricExportService => {
-    const responseRows: AnalysisResponseRow[] = [
-      {
-        responseId: 1,
-        personId: 1,
-        unitName: 'UNIT_A',
-        variableId: 'V1',
-        value: 'A',
-        codeV1: 1,
-        scoreV1: 1,
-        codeV2: 1,
-        scoreV2: 1,
-        codeV3: 1,
-        scoreV3: 1
-      },
-      {
-        responseId: 2,
-        personId: 2,
-        unitName: 'UNIT_A',
-        variableId: 'V1',
-        value: 'B',
-        codeV1: 0,
-        scoreV1: 0,
-        codeV2: 0,
-        scoreV2: 0,
-        codeV3: 0,
-        scoreV3: 0
-      },
-      {
-        responseId: 3,
-        personId: 3,
-        unitName: 'UNIT_A',
-        variableId: 'V1',
-        value: 'A',
-        codeV1: 1,
-        scoreV1: 1,
-        codeV2: 1,
-        scoreV2: 1,
-        codeV3: 1,
-        scoreV3: 1
-      },
-      {
-        responseId: 4,
-        personId: 4,
-        unitName: 'UNIT_A',
-        variableId: 'V1',
-        value: '[false,false,false,false]',
-        codeV1: 0,
-        scoreV1: 0,
-        codeV2: 0,
-        scoreV2: 0,
-        codeV3: 0,
-        scoreV3: 0
-      },
-      {
-        responseId: 5,
-        personId: 1,
-        unitName: 'UNIT_A',
-        variableId: 'HELPER',
-        value: 'x',
-        codeV1: 1,
-        scoreV1: 1,
-        codeV2: 1,
-        scoreV2: 1,
-        codeV3: 1,
-        scoreV3: 1
-      },
-      {
-        responseId: 6,
-        personId: 1,
-        unitName: 'UNIT_A',
-        variableId: 'HELPER',
-        value: 'y',
-        codeV1: 1,
-        scoreV1: 1,
-        codeV2: 1,
-        scoreV2: 1,
-        codeV3: 1,
-        scoreV3: 1
-      }
-    ].map(row => ({
-      statusV1: 5,
-      statusV2: 5,
-      statusV3: 5,
-      ...row,
-      ...responseOverrides[row.responseId]
-    }));
-    const responseRepository = {
-      createQueryBuilder: jest.fn(() => {
-        let grouped = false;
-        let lastResponseId = 0;
-        let statusKey: 'statusV1' | 'statusV2' | 'statusV3' | null = null;
-        let ignoredStatuses: number[] = [];
-        const getScopedRows = () => responseRows.filter(row => {
-          if (statusKey === null) {
-            return true;
-          }
-          const status = row[statusKey];
-          return status !== null && !ignoredStatuses.includes(status);
-        });
-        const queryBuilder = {
-          innerJoin: jest.fn().mockReturnThis(),
-          where: jest.fn().mockReturnThis(),
-          andWhere: jest.fn(
-            (
-              condition: string,
-              parameters?: {
-                lastResponseId?: number;
-                psychometricIgnoredStatuses?: number[];
-              }
-            ) => {
-              if (condition.includes('response.id >')) {
-                lastResponseId = parameters?.lastResponseId || 0;
-              }
-              const statusMatch = condition.match(
-                /response\.status_(v[123])/
-              );
-              if (statusMatch) {
-                statusKey =
-                  `status${statusMatch[1].toUpperCase()}` as
-                  'statusV1' | 'statusV2' | 'statusV3';
-              }
-              if (parameters?.psychometricIgnoredStatuses) {
-                ignoredStatuses =
-                  parameters.psychometricIgnoredStatuses;
-              }
-              return queryBuilder;
-            }
-          ),
-          select: jest.fn().mockReturnThis(),
-          addSelect: jest.fn().mockReturnThis(),
-          groupBy: jest.fn(() => {
-            grouped = true;
-            return queryBuilder;
-          }),
-          addGroupBy: jest.fn().mockReturnThis(),
-          having: jest.fn().mockReturnThis(),
-          orderBy: jest.fn().mockReturnThis(),
-          limit: jest.fn().mockReturnThis(),
-          getCount: jest.fn(async () => getScopedRows().length),
-          getRawMany: jest.fn(async () => {
-            if (grouped) {
-              const counts = new Map<string, number>();
-              getScopedRows().forEach(row => {
-                const key = [
-                  row.personId,
-                  row.unitName.toUpperCase(),
-                  row.variableId.toUpperCase()
-                ].join('\u001F');
-                counts.set(key, (counts.get(key) || 0) + 1);
-              });
-              return Array.from(counts.entries())
-                .filter(([, count]) => count > 1)
-                .map(([key]) => {
-                  const [personId, unitName, variableId] =
-                    key.split('\u001F');
-                  return {
-                    personId: Number(personId),
-                    unitName,
-                    variableId
-                  };
-                });
-            }
-            return getScopedRows().filter(
-              row => row.responseId > lastResponseId
-            );
-          })
-        };
-        return queryBuilder;
-      })
-    };
-    const queryRunner = {
-      manager: {
-        getRepository: jest.fn().mockReturnValue(responseRepository)
-      },
-      connect: jest.fn().mockResolvedValue(undefined),
-      startTransaction: jest.fn().mockResolvedValue(undefined),
-      query: jest.fn().mockResolvedValue([]),
-      commitTransaction: jest.fn().mockResolvedValue(undefined),
-      rollbackTransaction: jest.fn().mockResolvedValue(undefined),
-      release: jest.fn().mockResolvedValue(undefined),
-      isTransactionActive: true,
-      isReleased: false
-    };
-    const connection = {
-      createQueryRunner: jest.fn().mockReturnValue(queryRunner),
-      queryRunner
-    };
-
-    return new CodingPsychometricExportService(
-      {
-        find: jest.fn().mockResolvedValue([
-          {
-            id: 1,
-            file_id: 'UNIT_A.VOMD',
-            filename: 'UNIT_A.vomd',
-            data: JSON.stringify({
-              items: [
-                {
-                  id: 'ITEM_1',
-                  variableId: 'V1',
-                  description: '=Item label'
-                }
-              ]
-            })
-          }
-        ])
-      } as never,
-      {
-        getUnitVariableDetails: jest.fn().mockResolvedValue([
-          {
-            unitName: 'UNIT_A',
-            unitId: 'UNIT_A',
-            variables: [
-              {
-                id: 'V1',
-                alias: 'V1',
-                type: 'string',
-                multiple: true,
-                hasCodingScheme: true,
-                codes: [
-                  { id: 0, label: 'incorrect', score: 0 },
-                  { id: 1, label: 'correct', score: 1 }
-                ],
-                values: [
-                  { value: 'A', label: 'Option A' },
-                  { value: 'B', label: 'Option B' },
-                  { value: 'C', label: 'Unused option' },
-                  { value: '=2+2', label: '=Formula label' }
-                ]
-              },
-              {
-                id: 'HELPER',
-                alias: 'HELPER',
-                type: 'string',
-                hasCodingScheme: false
-              }
-            ]
-          }
-        ])
-      } as never,
-      {
-        resolveExclusionsForQueries: jest.fn().mockResolvedValue({
-          globalIgnoredUnits: [],
-          ignoredBooklets: [],
-          testletIgnoredUnits: []
-        })
-      } as never,
-      {
-        resolveMissingsProfileId: jest.fn().mockResolvedValue(1),
-        getMissingsProfileDetails: jest.fn().mockResolvedValue({
-          parseMissings: () => [
-            {
-              id: 'mir',
-              code: -98,
-              score: 0,
-              label: 'MIR'
-            },
-            {
-              id: 'mci',
-              code: -97,
-              score: null,
-              label: 'MCI'
-            }
-          ]
-        })
-      } as never,
-      connection as never
-    );
-  };
-
-  it('discovers complete, single-valued item domain fields from VOMD', async () => {
-    const service = createService(
-      {
-        items: [
-          {
-            id: 'I1',
-            variableId: 'V1',
-            profiles: [
-              {
-                profileId: 'profile',
-                entries: [
-                  {
-                    id: 'domain',
-                    label: [{ lang: 'de', value: 'Kompetenzbereich' }],
-                    value: [
-                      {
-                        id: 'D1',
-                        label: [{ lang: 'de', value: 'Domäne 1' }],
-                        annotation: []
-                      }
-                    ]
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            id: 'I2',
-            variableId: 'V2',
-            profiles: [
-              {
-                profileId: 'profile',
-                entries: [
-                  {
-                    id: 'domain',
-                    label: [{ lang: 'de', value: 'Kompetenzbereich' }],
-                    value: [
-                      {
-                        id: 'D2',
-                        label: [{ lang: 'de', value: 'Domäne 2' }],
-                        annotation: []
-                      }
-                    ]
-                  }
-                ]
-              }
-            ]
-          }
-        ]
-      },
-      [
-        {
-          id: 'V1',
-          alias: 'V1',
-          type: 'string',
-          hasCodingScheme: true
-        },
-        {
-          id: 'V2',
-          alias: 'V2',
-          type: 'string',
-          hasCodingScheme: true
-        }
-      ]
-    );
-
-    await expect(service.getDomainCandidates(7)).resolves.toEqual({
-      candidates: [
-        expect.objectContaining({
-          scope: 'ITEM',
-          profileId: 'profile',
-          entryId: 'domain',
-          label: 'Kompetenzbereich',
-          coverage: 2,
-          itemCount: 2,
-          singleValued: true,
-          selectable: true
-        })
-      ],
-      mappingIssueCount: 0
-    });
-  });
-
-  it('marks incomplete or multi-valued VOMD fields as unavailable', async () => {
-    const service = createService(
-      {
-        items: [
-          {
-            id: 'I1',
-            variableId: 'V1',
-            profiles: [
-              {
-                profileId: 'profile',
-                entries: [
-                  {
-                    id: 'domain',
-                    value: [{ id: 'D1' }, { id: 'D2' }]
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            id: 'I2',
-            variableId: 'V2',
-            profiles: []
-          }
-        ]
-      },
-      [
-        {
-          id: 'V1',
-          alias: 'V1',
-          type: 'string',
-          hasCodingScheme: true
-        },
-        {
-          id: 'V2',
-          alias: 'V2',
-          type: 'string',
-          hasCodingScheme: true
-        }
-      ]
-    );
-
-    const { candidates, mappingIssueCount } =
-      await service.getDomainCandidates(7);
-    const candidate = candidates[0];
-    expect(candidate).toEqual(
-      expect.objectContaining({
-        coverage: 1,
-        itemCount: 2,
-        singleValued: false,
-        selectable: false
-      })
-    );
-    expect(mappingIssueCount).toBe(0);
-  });
-
-  it('reports mapping issues when no domain candidates can be mapped', async () => {
-    const service = createService(
-      {
-        items: [
-          {
-            id: 'I1',
-            variableId: 'UNKNOWN'
-          }
-        ]
-      },
-      [
-        {
-          id: 'V1',
-          alias: 'V1',
-          type: 'string',
-          hasCodingScheme: true
-        }
-      ]
-    );
+  it('delegates domain candidate discovery to the metadata resolver', async () => {
+    const { service, metadataResolver } = createService();
 
     await expect(service.getDomainCandidates(7)).resolves.toEqual({
       candidates: [],
-      mappingIssueCount: 1
+      mappingIssueCount: 0
     });
+    expect(metadataResolver.getDomainCandidates).toHaveBeenCalledWith(7);
   });
 
-  it('ignores obsolete VOMD profile values', async () => {
-    const service = createService(
-      {
-        items: [
-          {
-            id: 'I1',
-            variableId: 'V1',
-            profiles: [
-              {
-                profileId: 'profile',
-                isCurrent: false,
-                entries: [{ id: 'old-domain', value: [{ id: 'OLD' }] }]
-              },
-              {
-                profileId: 'profile',
-                isCurrent: true,
-                entries: [{ id: 'domain', value: [{ id: 'D1' }] }]
-              }
-            ]
-          }
-        ]
-      },
-      [
-        {
-          id: 'V1',
-          alias: 'V1',
-          type: 'string',
-          hasCodingScheme: true
-        }
-      ]
+  it('orchestrates metadata, snapshot analysis, and Excel writing', async () => {
+    const {
+      service,
+      mapping,
+      snapshot,
+      analysis,
+      metadataResolver,
+      responseReader,
+      analysisEngine,
+      exportWriter
+    } = createService();
+
+    await service.writePsychometricsExcelToFile('/tmp/export.xlsx', {
+      workspaceId: 7,
+      version: 'v3',
+      domain: { mode: 'workspace' },
+      partWholeCorrection: false,
+      maxCategoryCount: 12
+    });
+
+    expect(metadataResolver.buildItemMapping).toHaveBeenCalledWith(7);
+    expect(metadataResolver.assignDomains).toHaveBeenCalledWith(
+      mapping,
+      { mode: 'workspace' }
     );
-
-    const { candidates } = await service.getDomainCandidates(7);
-    expect(candidates).toEqual([
-      expect.objectContaining({ entryId: 'domain', selectable: true })
-    ]);
-  });
-
-  it('uses selected missing codes and scores and parses multi-value categories', () => {
-    const service = createService({ items: [] }, []);
-    const missingDefinitions = [
+    expect(metadataResolver.loadMissingDefinitions).toHaveBeenCalledWith(
+      7,
+      undefined
+    );
+    expect(responseReader.withSnapshot).toHaveBeenCalledWith(
       {
-        id: 'mir',
-        code: -98,
-        score: 0,
-        label: 'MIR'
-      },
-      {
-        id: 'mci',
-        code: -97,
-        score: null,
-        label: 'MCI'
-      }
-    ];
-
-    expect(
-      (
-        service as never as {
-          normalizeCode: (
-            code: number,
-            definitions: unknown[]
-          ) => number | null;
-        }
-      ).normalizeCode(-3, missingDefinitions)
-    ).toBe(-98);
-    expect(
-      (
-        service as never as {
-          normalizeScore: (
-            value: { code: number; score: number | null },
-            code: number,
-            definitions: unknown[]
-          ) => number | null;
-        }
-      ).normalizeScore({ code: -3, score: null }, -98, missingDefinitions)
-    ).toBe(0);
-
-    const categories = (
-      service as never as {
-        parseCategories: (
-          value: string,
-          variable: Record<string, unknown>
-        ) => Array<{ value: string }>;
-      }
-    ).parseCategories('[true,false,true]', {
-      multiple: true,
-      values: [
-        { value: 'A', label: 'Kategorie A' },
-        { value: 'B', label: 'Kategorie B' },
-        { value: 'C', label: 'Kategorie C' }
-      ]
-    });
-    expect(categories.map(category => category.value)).toEqual(['A', 'C']);
-
-    const isCategoryEligible = (
-      service as never as {
-        isCategoryEligible: (code: number | null) => boolean;
-      }
-    ).isCategoryEligible.bind(service);
-    expect(isCategoryEligible(0)).toBe(true);
-    expect(isCategoryEligible(-97)).toBe(false);
-    expect(isCategoryEligible(null)).toBe(false);
-  });
-
-  it('keeps category discovery bounded by the configured limit', () => {
-    const service = createService({ items: [] }, []);
-    const item = {
-      variable: {
-        values: Array.from({ length: 1000 }, (_value, index) => ({
-          value: `category-${index}`,
-          label: `Category ${index}`
-        }))
-      },
-      codeDefinitions: new Map(),
-      categories: new Map(),
-      categoryLimitExceeded: false
-    };
-
-    (
-      service as never as {
-        initializeMetricDefinitions: (
-          items: unknown[],
-          missingDefinitions: unknown[],
-          maxCategoryCount: number
-        ) => void;
-      }
-    ).initializeMetricDefinitions([item], [], 10);
-
-    expect(item.categories.size).toBe(10);
-    expect(item.categoryLimitExceeded).toBe(true);
-  });
-
-  it('preserves explicitly valid empty response categories', () => {
-    const service = createService({ items: [] }, []);
-
-    const categories = (
-      service as never as {
-        parseCategories: (
-          value: string,
-          variable: Record<string, unknown>
-        ) => Array<{ value: string; label: string }>;
-      }
-    ).parseCategories('', {
-      processing: ['TAKE_EMPTY_AS_VALID'],
-      values: [{ value: '', label: 'Leere Antwort' }]
-    });
-
-    expect(categories).toEqual([
-      {
-        value: '___EMPTY___',
-        label: 'Leere Antwort',
-        source: 'UNIT_DEFINITION'
-      }
-    ]);
-  });
-
-  it('uses a domain mean and optionally removes the current item score', () => {
-    const service = createService({ items: [] }, []);
-    const getDomainScore = (
-      service as never as {
-        getDomainScore: (
-          aggregate: { sum: number; count: number },
-          itemScore: number | null,
-          partWholeCorrection: boolean
-        ) => number | null;
-      }
-    ).getDomainScore.bind(service);
-
-    expect(getDomainScore({ sum: 8, count: 3 }, 2, false)).toBeCloseTo(8 / 3);
-    expect(getDomainScore({ sum: 8, count: 3 }, 2, true)).toBe(3);
-    expect(getDomainScore({ sum: 2, count: 1 }, 2, true)).toBeNull();
-  });
-
-  it('maps VOMD items instead of requiring metadata for every unit variable', async () => {
-    const service = createAnalysisService();
-
-    const csv = await collectStream(
-      await service.exportPsychometricsAsCsv({
         workspaceId: 7,
-        version: 'v2',
+        version: 'v3',
+        mapping
+      },
+      expect.any(Function),
+      undefined
+    );
+    expect(analysisEngine.analyze).toHaveBeenCalledWith({
+      options: expect.objectContaining({
+        workspaceId: 7,
+        version: 'v3',
         partWholeCorrection: false,
-        domain: { mode: 'workspace' }
-      })
+        maxCategoryCount: 12
+      }),
+      mapping,
+      missingDefinitions: [],
+      snapshot
+    });
+    expect(exportWriter.writeExcelToFile).toHaveBeenCalledWith(
+      '/tmp/export.xlsx',
+      analysis,
+      undefined
     );
-
-    expect(csv).toContain('SCORE;WORKSPACE');
-    expect(csv).toContain('CATEGORY;WORKSPACE');
-    expect(csv).not.toContain(';HELPER;');
   });
 
-  it('excludes codes and scores with ignored statuses for the selected version', async () => {
-    const overrides = {
-      1: {
-        statusV2: 0,
-        codeV2: 99,
-        scoreV2: 99
-      }
-    };
-    const v2Service = createAnalysisService(overrides);
-    const v2Analysis = await (
-      v2Service as never as {
-        analyze: (options: {
-          workspaceId: number;
-          version: 'v2';
-          partWholeCorrection: boolean;
-        }) => Promise<{
-          rows: Array<{
-            type: string;
-            code: string;
-            n: number;
-          }>;
-        }>;
-      }
-    ).analyze({
+  it('passes CSV output cancellation into the analysis pipeline', async () => {
+    const {
+      service,
+      analysis,
+      analysisEngine,
+      exportWriter
+    } = createService();
+    const jobCancellation = jest.fn().mockResolvedValue(undefined);
+    const outputCancellation = jest.fn().mockResolvedValue(undefined);
+
+    await service.exportPsychometricsAsCsv({
       workspaceId: 7,
-      version: 'v2',
-      partWholeCorrection: false
+      checkCancellation: jobCancellation
     });
+    const analyze = exportWriter.createCsvStream.mock.calls[0][0] as (
+      checkCancellation: () => Promise<void>
+    ) => Promise<PsychometricAnalysis>;
 
-    expect(
-      v2Analysis.rows.find(row => row.type === 'SCORE')
-    ).toEqual(expect.objectContaining({ n: 3 }));
-    expect(
-      v2Analysis.rows.some(
-        row => row.type === 'CODE' && row.code === '99'
-      )
-    ).toBe(false);
-
-    const v1Service = createAnalysisService(overrides);
-    const v1Analysis = await (
-      v1Service as never as {
-        analyze: (options: {
-          workspaceId: number;
-          version: 'v1';
-          partWholeCorrection: boolean;
-        }) => Promise<{
-          rows: Array<{
-            type: string;
-            n: number;
-          }>;
-        }>;
-      }
-    ).analyze({
-      workspaceId: 7,
-      version: 'v1',
-      partWholeCorrection: false
-    });
-
-    expect(
-      v1Analysis.rows.find(row => row.type === 'SCORE')
-    ).toEqual(expect.objectContaining({ n: 4 }));
-  });
-
-  it('uses one read-only repeatable-read snapshot for both response passes', async () => {
-    const service = createAnalysisService();
-    const dependencies = service as never as {
-      connection: {
-        queryRunner: {
-          startTransaction: jest.Mock;
-          query: jest.Mock;
-          commitTransaction: jest.Mock;
-          rollbackTransaction: jest.Mock;
-          release: jest.Mock;
-        };
-      };
-      workspaceExclusionService: {
-        resolveExclusionsForQueries: jest.Mock;
-      };
-    };
-
-    await collectStream(
-      await service.exportPsychometricsAsCsv({
-        workspaceId: 7,
-        partWholeCorrection: false
-      })
+    await expect(analyze(outputCancellation)).resolves.toBe(analysis);
+    expect(exportWriter.createCsvStream).toHaveBeenCalledWith(
+      expect.any(Function),
+      jobCancellation
     );
-
-    expect(
-      dependencies.workspaceExclusionService.resolveExclusionsForQueries
-    ).toHaveBeenCalledTimes(1);
-    expect(
-      dependencies.connection.queryRunner.startTransaction
-    ).toHaveBeenCalledWith('REPEATABLE READ');
-    expect(dependencies.connection.queryRunner.query).toHaveBeenCalledWith(
-      'SET TRANSACTION READ ONLY'
-    );
-    expect(
-      dependencies.connection.queryRunner.commitTransaction
-    ).toHaveBeenCalledTimes(1);
-    expect(
-      dependencies.connection.queryRunner.rollbackTransaction
-    ).not.toHaveBeenCalled();
-    expect(dependencies.connection.queryRunner.release).toHaveBeenCalledTimes(
-      1
-    );
-  });
-
-  it('returns the CSV stream before analysis completes', async () => {
-    const service = createAnalysisService();
-    let resolveAnalysis:
-    ((analysis: { rows: []; summary: [] }) => void) | undefined;
-    const analysisPromise = new Promise<{
-      rows: [];
-      summary: [];
-    }>(resolve => {
-      resolveAnalysis = resolve;
-    });
-    jest
-      .spyOn(service as never as { analyze: () => Promise<unknown> }, 'analyze')
-      .mockReturnValue(analysisPromise);
-
-    let returnedStream: NodeJS.ReadableStream | undefined;
-    service
-      .exportPsychometricsAsCsv({
-        workspaceId: 7,
-        partWholeCorrection: false
-      })
-      .then(stream => {
-        returnedStream = stream;
-      });
-
-    await new Promise<void>(resolve => {
-      setImmediate(resolve);
-    });
-    expect(returnedStream).toBeDefined();
-
-    const csvPromise = collectStream(returnedStream!);
-    resolveAnalysis?.({ rows: [], summary: [] });
-
-    await expect(csvPromise).resolves.toContain('type;domain;domain_label');
-  });
-
-  it('stops CSV production when the output stream is destroyed', async () => {
-    const service = createAnalysisService();
-    let resolveAnalysis:
-    ((analysis: { rows: unknown[]; summary: [] }) => void) | undefined;
-    const analysisPromise = new Promise<{
-      rows: unknown[];
-      summary: [];
-    }>(resolve => {
-      resolveAnalysis = resolve;
-    });
-    jest
-      .spyOn(service as never as { analyze: () => Promise<unknown> }, 'analyze')
-      .mockReturnValue(analysisPromise);
-    const writer = jest.spyOn(
-      service as never as {
-        writePsychometricsCsv: (
-          output: NodeJS.WritableStream,
-          options: unknown
-        ) => Promise<void>;
-      },
-      'writePsychometricsCsv'
-    );
-
-    const stream = await service.exportPsychometricsAsCsv({
-      workspaceId: 7,
-      partWholeCorrection: false
-    });
-    stream.on('error', () => undefined);
-    const closed = new Promise<void>(resolve => {
-      stream.once('close', resolve);
-    });
-    (
-      stream as NodeJS.ReadableStream & {
-        destroy: (error?: Error) => void;
-      }
-    ).destroy(new Error('destination failed'));
-    await closed;
-
-    const row = {
-      type: 'SCORE',
-      domain: 'WORKSPACE',
-      domainLabel: 'Gesamter Workspace',
-      unit: 'UNIT_A',
-      item: 'I1',
-      variable: 'V1',
-      itemLabel: 'x'.repeat(500),
-      code: '',
-      category: '',
-      label: '',
-      score: '',
-      source: '',
-      n: 1,
-      positiveN: '',
-      positiveShare: '',
-      correlation: '',
-      status: 'TOO_FEW_CASES',
-      note: ''
-    };
-    resolveAnalysis?.({
-      rows: Array.from({ length: 2000 }, () => row),
-      summary: []
-    });
-
-    const production = writer.mock.results[0].value;
-    let timeout: ReturnType<typeof setTimeout> | undefined;
-    const completion = Promise.race([
-      production,
-      new Promise<never>((_, reject) => {
-        timeout = setTimeout(
-          () => reject(new Error('CSV production did not stop')),
-          500
-        );
-      })
-    ]);
-
-    await expect(completion).resolves.toBeUndefined();
-    if (timeout) {
-      clearTimeout(timeout);
-    }
-  });
-
-  it('propagates output destruction into the running CSV analysis', async () => {
-    const service = createAnalysisService();
-    let continueAnalysis: (() => void) | undefined;
-    let analysisStarted: (() => void) | undefined;
-    const analysisStartedPromise = new Promise<void>(resolve => {
-      analysisStarted = resolve;
-    });
-    const continueAnalysisPromise = new Promise<void>(resolve => {
-      continueAnalysis = resolve;
-    });
-    jest
-      .spyOn(
-        service as never as {
-          analyze: (options: {
-            checkCancellation?: () => Promise<void>;
-          }) => Promise<unknown>;
-        },
-        'analyze'
-      )
-      .mockImplementation(async options => {
-        analysisStarted?.();
-        await continueAnalysisPromise;
-        await options.checkCancellation?.();
-        return { rows: [], summary: [] };
-      });
-    const writer = jest.spyOn(
-      service as never as {
-        writePsychometricsCsv: (
-          output: NodeJS.WritableStream,
-          options: unknown
-        ) => Promise<void>;
-      },
-      'writePsychometricsCsv'
-    );
-
-    const stream = await service.exportPsychometricsAsCsv({
-      workspaceId: 7,
-      partWholeCorrection: false
-    });
-    await analysisStartedPromise;
-    stream.on('error', () => undefined);
-    const closed = new Promise<void>(resolve => {
-      stream.once('close', resolve);
-    });
-    (
-      stream as NodeJS.ReadableStream & {
-        destroy: (error?: Error) => void;
-      }
-    ).destroy(new Error('destination failed'));
-    await closed;
-
-    continueAnalysis?.();
-
-    await expect(writer.mock.results[0].value).resolves.toBeUndefined();
-  });
-
-  it('destroys the Excel output stream when finalization is cancelled', async () => {
-    const service = createAnalysisService();
-    jest
-      .spyOn(service as never as { analyze: () => Promise<unknown> }, 'analyze')
-      .mockResolvedValue({ rows: [], summary: [] });
-    const outputStream = new PassThrough();
-    const createWriteStream = jest
-      .spyOn(fs, 'createWriteStream')
-      .mockReturnValue(outputStream as never);
-    const cancellationError = new Error('cancelled');
-    let cancellationChecks = 0;
-
-    try {
-      await expect(
-        service.writePsychometricsExcelToFile('/tmp/psychometrics.xlsx', {
-          workspaceId: 7,
-          checkCancellation: async () => {
-            cancellationChecks += 1;
-            if (cancellationChecks >= 2) {
-              throw cancellationError;
-            }
-          }
-        })
-      ).rejects.toThrow(cancellationError);
-
-      expect(outputStream.destroyed).toBe(true);
-    } finally {
-      createWriteStream.mockRestore();
-    }
-  });
-
-  it('exports zero dummies for unused categories and empty multiple responses', async () => {
-    const service = createAnalysisService();
-    const analysis = await (
-      service as never as {
-        analyze: (options: {
-          workspaceId: number;
-          partWholeCorrection: boolean;
-        }) => Promise<{
-          rows: Array<{
-            type: string;
-            category: string;
-            n: number;
-            positiveN: number | '';
-            status: string;
-          }>;
-        }>;
-      }
-    ).analyze({ workspaceId: 7, partWholeCorrection: false });
-
-    expect(analysis.rows).toContainEqual(
+    expect(analysisEngine.analyze).toHaveBeenCalledWith(
       expect.objectContaining({
-        type: 'CATEGORY',
-        category: 'C',
-        n: 4,
-        positiveN: 0,
-        status: 'CONSTANT_ITEM'
+        options: expect.objectContaining({
+          checkCancellation: outputCancellation
+        })
       })
-    );
-  });
-
-  it('sanitizes formula-like text in psychometric CSV exports', async () => {
-    const service = createAnalysisService();
-
-    const csv = await collectStream(
-      await service.exportPsychometricsAsCsv({
-        workspaceId: 7,
-        partWholeCorrection: false
-      })
-    );
-
-    expect(csv).toContain("'=Item label");
-    expect(csv).toContain("'=2+2");
-    expect(csv).toContain("'=Formula label");
-    expect(csv).not.toMatch(/(^|;)=2\+2(;|$)/m);
-  });
-
-  it('reports VOMD items whose variableId cannot be resolved', async () => {
-    const service = createService(
-      {
-        items: [
-          {
-            id: 'ITEM_MISSING',
-            variableId: 'UNKNOWN'
-          }
-        ]
-      },
-      [
-        {
-          id: 'V1',
-          alias: 'V1',
-          type: 'string',
-          hasCodingScheme: true
-        }
-      ]
-    );
-
-    const mapping = await (
-      service as never as {
-        buildItemMappingContext: (workspaceId: number) => Promise<{
-          items: unknown[];
-          issues: string[];
-        }>;
-      }
-    ).buildItemMappingContext(7);
-
-    expect(mapping.items).toHaveLength(0);
-    expect(mapping.issues).toContain(
-      'UNIT_A/ITEM_MISSING: Variable UNKNOWN nicht gefunden'
     );
   });
 });
